+++ src/child.c
@@ -278,6 +278,22 @@ trim_environment(void)
   trimenv("BYOBU_");
 }
 
+static bool
+ispathprefix(string pref, string path)
+{
+  if (*pref == '/')
+    pref++;
+  if (*path == '/')
+    path++;
+  int len = strlen(pref);
+  if (0 == strncmp(pref, path, len)) {
+    path += len;
+    if (!*path || *path == '/')
+      return true;
+  }
+  return false;
+}
+
 void
 child_create(char *argv[], struct winsize *winp)
 {
@@ -293,6 +309,19 @@ child_create(char *argv[], struct winsize *winp)
   signal(SIGINT, sigexit);
   signal(SIGTERM, sigexit);
   signal(SIGQUIT, sigexit);
+
+  // support OSC 7 directory cloning if cloning WSL window while in rootfs
+  if (support_wsl && wslname) {
+    // this is done once in a new window, so don't care about memory leaks
+    char * rootfs = path_win_w_to_posix(wsl_basepath);
+    char * cwd = getcwd(malloc(MAX_PATH), MAX_PATH);
+    if (ispathprefix(rootfs, cwd)) {
+      char * dist = cs__wcstombs(wslname);
+      char * wslsubdir = cwd + strlen(rootfs);
+      char * wsldir = asform("//wsl$/%s%s", dist, wslsubdir);
+      chdir(wsldir);
+    }
+  }
 
   // Create the child process and pseudo terminal.
   pid = forkpty(&pty_fd, 0, 0, winp);